+++ build_runner/lib/src/build_plan/builder_definition.dart
@@ -129,6 +129,9 @@ class BuilderDefinition implements AbstractBuilderDefinition {
   /// Whether the builder is skipped if nothing uses its output.
   final bool isOptional;
 
+  /// Whether the builder is capable of contributing to .part files.
+  final bool addsToLibrary;
+
   @visibleForTesting
   BuilderDefinition(
     this.key, {
@@ -137,6 +140,7 @@ class BuilderDefinition implements AbstractBuilderDefinition {
     Iterable<String> appliesBuilders = const [],
     this.hideOutput = true,
     this.isOptional = false,
+    this.addsToLibrary = false,
     this.targetBuilderConfigDefaults = const TargetBuilderConfigDefaults(),
   }) : package = package ?? (key.contains(':') ? key.split(':').first : ''),
        appliesBuilders = appliesBuilders.toBuiltList();
@@ -150,6 +154,7 @@ class BuilderDefinition implements AbstractBuilderDefinition {
     appliesBuilders: builderDefinition.appliesBuilders,
     hideOutput: builderDefinition.buildTo == build_config.BuildTo.cache,
     isOptional: builderDefinition.isOptional,
+    addsToLibrary: builderDefinition.addsToLibrary,
     targetBuilderConfigDefaults: builderDefinition.defaults,
   );